+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/application/CircuitBreakerFactory.java
@@ -2,7 +2,6 @@ package com.mageddo.dnsproxyserver.solver.remote.application;
 
 import com.mageddo.commons.circuitbreaker.CircuitCheckException;
 import com.mageddo.dnsproxyserver.config.application.ConfigService;
-import com.mageddo.dnsproxyserver.solver.application.CircuitBreakerStateMapper;
 import com.mageddo.dnsproxyserver.solver.remote.Result;
 import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAO;
 import dev.failsafe.CircuitBreaker;
